+++ daemon/src/config.rs
@@ -17,6 +17,7 @@ pub struct Config {
     pub gfx_last_mode: GfxVendors,
     pub gfx_managed: bool,
     pub gfx_vfio_enable: bool,
+    pub gfx_save_compute_vfio: bool,
     pub active_profile: String,
     pub toggle_profiles: Vec<String>,
     #[serde(skip)]
@@ -37,6 +38,7 @@ impl Default for Config {
             gfx_last_mode: GfxVendors::Hybrid,
             gfx_managed: true,
             gfx_vfio_enable: false,
+            gfx_save_compute_vfio: true,
             active_profile: "normal".into(),
             toggle_profiles: vec!["normal".into(), "boost".into(), "silent".into()],
             curr_fan_mode: 0,