+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+# [3.5.0] - 2021-04-25
 ### Changed
 + Keyboard:
   - Split out all aura functionality that isn't dependent on the daemon in to a
-    new crate `rog-aura`
+    new crate `rog-aura` (incomplete)
   - Keyboard LED control now includes:
     + Enable/disable LED's while laptop is awake
     + Enable/disable LED animation while laptop is suspended and AC plugged in